requirements.txt
azure-functions
azure-functions-durable
requests
pylint